Split flattener in separate parts
